pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SALTISE_course_flow"
version = "0.8.21"
description = "CourseFlow"
readme = "README.rst"
authors = [
    { name = "Arthur Ayestas Hilgert", email = "aahilgert@gmail.com" },
    { name = "Jeremie Choquette" }
]
license = { file = "LICENSE.txt" }
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
homepage = "https://www.example.com/"

[project.optional-dependencies]
default = [
    "asgiref",
    "bleach",
    "celery[redis]",
    "channels >= 4",
    "channels_redis >= 4",
    "django >= 3, < 4",
    "django-compressor",
    "django-csp",
    "django-model-utils",
    "django-ratelimit < 4",
    "django-redis",
    "djangorestframework",
    "html2text",
    "openpyxl",
    "pandas >= 1.3.5",
    "xlrd",
    "xlsxwriter"
]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py39-django3_2

[testenv]
setenv =
  PIP_INDEX_URL = https://pypi.python.org/simple
deps =
  daphne >= 4
  pytest-django
  redislite
  selenium
allowlist_externals = sh
#commands_pre = sh -c 'redis-server &'
#commands = python -X dev -m pytest {posargs} --ignore-glob=*functional*
#commands_post = redis-cli shutdown
"""

[tool.black]
line_length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
/)
'''

[tool.isort]
skip = ["./course_flow/models/__init__.py"]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
